for #2408, move SQLStatementRuleDefinitionEntityLoaderTest
